usb: dwc3: Add support for masking phy reset signal
Versal usb controller drives phy reset signal during
Soft Reset or Core Reset to issue reset to the ULPI
phy. The ULPI phy needs to be resetted only at the time
of initialization. This patch updates the code to provide
a function for masking the phy reset signal from reaching
the ULPI phy. This function is called only after the phy
initialization is done.

Signed-off-by: Anurag Kumar Vulisha <anurag.kumar.vulisha@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
State: pending