media: i2c: max9286: Make the overlap window as control
The overlap window seems like an additional validation by tracking
a delay between vsync and frame sync: error if the delay is bigger than
the programmed window. It's not clear what value it should be as well.
Hence it can be optionally enabled.
Disable the overlap by default, but let user enable through control
if additional check is needed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>