dmaengine: xilinx: frmbuf: Support more compatible formats
Some formats are not supported, but not different with supported
formats from DMA point of view. An example is, VYUY and UYVY that are
unpacked identically. But VYUY is marked as supported, while UYVY is
not.

This patch adds a function to convert the unsupported but compatible
fourcc into supported one. For now, the conversion happens only when
the format verification fails, and the function has only single mapping
that is tested. But for sure there are more, so keeping it as a table
to extend easily.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>